--- app/config/full_text_search_config.go
@@ -1,13 +1,0 @@
-package config
-
-func OpenSearchURL() string {
-	return config.String("open.search.url")
-}
-
-func OpenSearchUsername() string {
-	return config.String("open.search.username")
-}
-
-func OpenSearchPassword() string {
-	return config.String("open.search.password")
-}
